Update DoctrineBundle version and enable profiling backtrace
Upgraded DoctrineBundle version from 2.10 to 2.11 in the symfony.lock file to incorporate the latest features and fixes of the bundle, ensuring up-to-date functionality.

Also, profiling_collect_backtrace option is enabled in the doctrine.yaml configuration to collect backtrace when in debug mode. This can provide more detailed debug information, helping improve the debugging process when necessary.