Upgrade PHPUnit to 9.6 and Symfony PHPUnit-bridge to 6.4
In this commit, upgraded PHPUnit version from 9.5 to 9.6 in 'phpunit.xml.dist' to benefit from new features and improvements in the latest PHPUnit version. Also, updated the version of symfony/phpunit-bridge from 6.3 to 6.4 in 'symfony.lock' for enhanced stability and resolution to known bugs. Removed an unnecessary line in 'phpunit.xml.dist' for cleaner code format.