Remove NelmioApiDocBundle and related settings
Removed usage of nelmio/api-doc-bundle which was being used for API documentation. This decision was made because the project requirements changed and the API documentation feature was no longer needed. As a result, the associated routes, bundle registration, and dependencies in composer.json were also removed. Configuration settings related to annotation processing in framework.yaml were also changed and annotation processing by Symfony is disabled.